+++ pkgs/intl4x/lib/src/ecma/ecma_stub.dart
@@ -1,3 +1,3 @@
-import '../ecma_policy.dart';
+import 'ecma_policy.dart';
 
 const EcmaPolicy defaultPolicy = NeverEcma();
